+++ .gitignore
@@ -7,5 +7,5 @@ build/*
 .externalNativeBuild
 .cxx
 local.properties
-.lock
-.bin
+*.lock
+*.bin
