+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.RepairTool">
         <activity
-            android:name=".volunteer.RepairActivity"
+            android:name=".volunteer.VolunteerActivity"
             android:exported="true"
             android:label="@string/title_activity_repair"
             android:theme="@style/Theme.RepairTool.NoActionBar" />
